+++ components/com_kinoarhiv/views/movie/view.html.php
@@ -37,6 +37,8 @@ class KinoarhivViewMovie extends JViewLegacy
 
 	protected $user;
 
+	protected $itemid;
+
 	/**
 	 * Execute and display a template script.
 	 *